add get endpoint to fetch a profile linked to an subdomain
