+++ packages/offchain-resolver/src/index.ts
@@ -4,7 +4,7 @@ import http from 'http';
 import cors from 'cors';
 import winston from 'winston';
 import { getDatabase } from './persistance/getDatabase';
-import { addProfileResource } from './http/addProfileResource';
+import { ccipGateway } from './http/ccipGateway';
 
 const app = express();
 app.use(express.json({ limit: '50mb' }));
@@ -22,5 +22,5 @@ app.use(bodyParser.json());
 
     app.locals.db = await getDatabase();
 
-    app.use(addProfileResource());
+    app.use(ccipGateway());
 })();
